add the auth and login

+++ app/Http/Controllers/ArtistController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Artist;
 use App\Models\Song;
-
+use Illuminate\Support\Facades\Gate;
 
 class ArtistController extends Controller
 {
@@ -25,6 +25,8 @@ class ArtistController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', Artist::class);
+
         return view('artists.create');
     }
 
@@ -33,11 +35,18 @@ class ArtistController extends Controller
      */
     public function store(Request $request) // Dependency Injection
     {
+        Gate::authorize('create', Artist::class);
+        
+        $request->validate([
+            'name' => ['required', 'string', 'min:3', 'max:255']
+        ]);
+
         $artist_name = $request->get('name');
+
         if($artist_name == null){
             return redirect()->back();
         }
-        
+
         $artist = new Artist();
         $artist->name = $artist_name;
         $artist->save();
@@ -60,6 +69,7 @@ class ArtistController extends Controller
      */
     public function edit(Artist $artist)
     {
+        Gate::authorize('update', $artist);
         return view('artists.edit',[
             'artist' => $artist
         ]);
@@ -70,6 +80,11 @@ class ArtistController extends Controller
      */
     public function update(Request $request, Artist $artist)
     {
+        Gate::authorize('update', $artist);
+        $request->validate([
+            'name' => ['required', 'string', 'min:4', 'max:255']
+        ]);
+
         $artist->name = $request->get('name');
         $artist->save();
         return redirect()->route('artists.show', ['artist' => $artist]);
@@ -80,6 +95,8 @@ class ArtistController extends Controller
      */
     public function destroy(Artist $artist)
     {
+        Gate::authorize('destroy', $artist);
+
         if($artist->songs->isEmpty()){
             $artist->delete();
             return redirect()->route('artists.index');
@@ -94,6 +111,8 @@ class ArtistController extends Controller
      * Route Name: artists.songs.create
      */
     public function createSong(Artist $artist){
+        Gate::authorize('update',$artist);
+
         return view('artists.create-song',[
             'artist' => $artist
         ]);
@@ -106,6 +125,13 @@ class ArtistController extends Controller
      * Route Name: artists.songs.store
      */
     public function storeSong(Request $request, Artist $artist){
+        Gate::authorize('update', $artist);
+        // ['required','string','min:4', 'max:255']
+        $request->validate([
+            'title' => 'required|string|min:4|max:255',
+            'duration' => ['required', 'integer', 'min:10']
+        ]);
+
         $song = new Song(); // use App\Models\Song;
         $song->title = $request->get('title');
         $song->duration = $request->get('duration');
